+++ app/views/home/index.html.erb
@@ -42,7 +42,7 @@
     </div>
 
     <div id="analitic" style="display:none;">
-      <%= form_tag analitics_path(format: 'xlsx') do  %>
+      <%= form_tag analitics_path do  %>
         <div class="field">
           <%= label_tag :email %>
           <%= text_field_tag :email %>
