+++ app/views/layouts/mailer.xlsx.axlsx
@@ -1,0 +1,1 @@
+yield
