+++ app/views/linkedin_mailer/followers.xlsx.axlsx
@@ -1,8 +1,8 @@
-@wb = xlsx_package.workbook
+wb = xlsx_package.workbook
 
-@wb.add_worksheet(name: 'Sheet') do |sheet|
+wb.add_worksheet(name: 'Sheet') do |sheet|
   if result.first
-    sheet.add_row @result.first.keys
+    sheet.add_row result.first.keys
     result.each do |row|
       sheet.add_row row.values
     end
